Ensure times and durations with fractional seconds and microsecond or lower resolution are not subject to rounding.
With the release of 6.0.0, the fractional components of seconds are cast
to floats before being truncated. These fractional components can be
cast directly to microseconds by left justifying with 0s, and truncating
to 6 characters before casting to an int. This prevents roundoff issues,
such as those mentioned in #24.

This does not fix rounding issues with fractional years, months, weeks,
days, or minutes.

This fixes #24.